requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
django-adminactions==1.13.1
black==21.8b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.15
django-compat==1.0.15
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-model-utils==4.1.1
django-online-users==0.3
django-redis==5.0.0
django-simple-history==3.0.0
django-taggit==1.5.1
djangorestframework==3.12.4
gunicorn==20.1.0
humanize==3.11.0
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
Markdown==3.3.4
markdown2==2.4.2
martor==1.6.4
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.3.0
pluggy==1.0.0
portalocker==2.3.2
psycopg2==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.5
python-slugify==5.0.2
pytz==2021.1
redis==3.5.3
regex==2021.8.28
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
yapf==0.31.0
zipp==3.5.0
isort